refactor formatter into cpp
