+++ rosbag2_py/CMakeLists.txt
@@ -79,6 +79,7 @@ clean_windows_flags(_reader)
 
 pybind11_add_module(_storage SHARED
   src/rosbag2_py/_storage.cpp
+  src/rosbag2_py/format_bag_metadata.cpp
 )
 ament_target_dependencies(_storage PUBLIC
   "rosbag2_cpp"
